game: retire -Wall warnings (offset-macro intptr casts, dead locals, dangling-else braces, switch defaults)

+++ game/local.h
@@ -20,6 +20,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 // local.h -- local definitions for game module
 
 #include "game/q_shared.h"
+#include <stdint.h>
 
 // define GAME_INCLUDE so that game.h does not define the
 // short, server-visible gclient_t and edict_t structures,
@@ -501,10 +502,10 @@ extern	int	meansOfDeath;
 
 extern	edict_t			*g_edicts;
 
-#define	FOFS(x) (int)&(((edict_t *)0)->x)
-#define	STOFS(x) (int)&(((spawn_temp_t *)0)->x)
-#define	LLOFS(x) (int)&(((level_locals_t *)0)->x)
-#define	CLOFS(x) (int)&(((gclient_t *)0)->x)
+#define	FOFS(x) (int)(intptr_t)&(((edict_t *)0)->x)
+#define	STOFS(x) (int)(intptr_t)&(((spawn_temp_t *)0)->x)
+#define	LLOFS(x) (int)(intptr_t)&(((level_locals_t *)0)->x)
+#define	CLOFS(x) (int)(intptr_t)&(((gclient_t *)0)->x)
 
 #define random()	((rand () & 0x7fff) / ((float)0x7fff))
 #define crandom()	(2.0 * (random() - 0.5))
